rename deepClone II
